Fix: Creating edge-commit to update the ostree repo will now refer to the existing ostree commit rather than building a independent commit (#480)
* Whenever we initiate a edge-commit to update existing ostree repo, its
not referring to the existing ostree repo as a parent commit but instead
builds an independent commit.
Due to this when we update rpm-ostree system using this repo, problems
have been observed with regards to UID of the system users which are
added by the rpm scriptlet since these are stored in /usr/lib/passwd
which changes across the commit since the commit are now not related to
each other.

This change aims to solve this problem by the follow changes,

 * Change the "if" condition in start_compose.py so that the parameters
   "ostree_url" and "ostree_ref" are not ignored for edge-commit

 * Added parameters "ostree_url" and "ostree_ref" to the builder role
   "Start compose" task such that they can be used in when updating
   existing commit and can be ignored when building the first commit

* Correction to the previous commit as per PR comments.
Removing the "# noqa yaml[line-length]" since the line is shorter and
does not need this.

* Changes to the previous commit as per PR comments.
Added ansible change log fragment.
Closes #477

* Update changelogs/fragments/313-edge-commit-should-update-existing-ostee-repo.yaml